refactor: move pena players query slice into core application
