+++ backend/src/core/application/ports/__init__.py
@@ -1,4 +1,5 @@
 from .nationality_query_port import NationalityQueryPort
+from .pena_player_query_port import PenaPlayerQueryPort
 from .pena_profile_port import PenaProfilePort
 from .pena_query_port import PenaQueryPort
 from .player_profile_port import PlayerProfilePort
@@ -6,6 +7,7 @@ from .season_match_insights_port import SeasonMatchInsightsPort
 
 __all__ = [
     "NationalityQueryPort",
+    "PenaPlayerQueryPort",
     "PenaProfilePort",
     "PenaQueryPort",
     "PlayerProfilePort",
