+++ backend/src/core/application/use_cases/__init__.py
@@ -1,4 +1,5 @@
 from .get_nationalities_usecase import GetNationalitiesUseCase
+from .get_pena_players_usecase import GetPenaPlayersUseCase
 from .get_penas_usecase import GetPenasUseCase
 from .get_player_profile_usecase import GetPlayerProfileUseCase
 from .get_season_match_insights_usecase import GetSeasonMatchInsightsUseCase
@@ -23,6 +24,7 @@ from .update_player_profile_usecase import (
 
 __all__ = [
     "GetNationalitiesUseCase",
+    "GetPenaPlayersUseCase",
     "GetPenasUseCase",
     "GetPlayerProfileUseCase",
     "GetSeasonMatchInsightsUseCase",
