+++ backend/tests/test_pena_players_controller.py
@@ -6,12 +6,9 @@ from api.interface.controller.v1.model.request.pena_players_request import (
 )
 from auth.dependencies import require_user
 from auth.session import SessionData
+from core.application.models import PenaPlayerInfo, PenaPlayersPage
 from fastapi import HTTPException
 from persistence.application.update_policies import FieldUpdate
-from persistence.application.use_cases.get_pena_players_usecase import (
-    PenaPlayerInfo,
-    PenaPlayersPage,
-)
 from persistence.application.use_cases.manage_pena_membership_usecase import (
     InvalidPenaGuestPlayerDataError,
     InvalidPenaMembershipUpdateDataError,
